Expandable server buffer + M1 testing + improved record string conversion (#13)
---------

+++ DnsCore/DnsDefaults.cs
@@ -3,5 +3,5 @@ namespace DnsCore;
 public static class DnsDefaults
 {
     public const ushort Port = 53;
-    public const int MaxUdpMessageSize = 512;
+    public const ushort MaxUdpMessageSize = 512;
 }
